+++ src/js/binaryen.js-post.js
@@ -3128,8 +3128,8 @@ Module['getExpressionInfo'] = function(expr) {
 };
 
 // Gets the side effects of the specified expression
-Module['getSideEffects'] = function(expr, features) {
-  return Module['_BinaryenExpressionGetSideEffects'](expr, features);
+Module['getSideEffects'] = function(expr, module) {
+  return Module['_BinaryenExpressionGetSideEffects'](expr, module);
 };
 
 Module['createType'] = function(types) {