+++ src/wasm-binary.h
@@ -1243,6 +1243,7 @@ public:
   void writeInlineString(const char* name);
   void writeEscapedName(const char* name);
   void writeInlineBuffer(const char* data, size_t size);
+  void writeData(const char* data, size_t size);
 
   struct Buffer {
     const char* data;
@@ -1251,12 +1252,6 @@ public:
     Buffer(const char* data, size_t size, size_t pointerLocation)
       : data(data), size(size), pointerLocation(pointerLocation) {}
   };
-
-  std::vector<Buffer> buffersToWrite;
-
-  void emitBuffer(const char* data, size_t size);
-  void emitString(const char* str);
-  void finishUp();
 
   Module* getModule() { return wasm; }