+++ CHANGELOG.md
@@ -24,6 +24,9 @@ Current Trunk
 - Adds type argument for BinaryenAddTable method (#4107). For the binaryen.js api
   this parameter is optional and by default is set to funcref type.
 
+- Replace `BinaryenExpressionGetSideEffects`'s features parameter with a module
+  parameter.
+
 v101
 ----